+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Bug fixes
 - Fix `Reacher-v3` name in PPO hyperparameter file
 - Pinned ale-py==0.7.4 until new SB3 version is released
+- Fix enjoy / record videos with LSTM policy
 
 ### Documentation