+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 ### Bug fixes
 - Policies saved during during optimization with distributed Optuna load on new systems (@jkterry)
 - Fixed script for recording video that was not up to date with the enjoy script
+- Added CEM support
 
 ### Documentation