+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### New Features
 - Support setting PyTorch's device with thye `--device` flag (@gregwar)
 - Added `vec_env_wrapper` support in the config (works the same as `env_wrapper`)
+- Added Huggingface hub integration
 
 ### Bug fixes
 - Fix `Reacher-v3` name in PPO hyperparameter file