+++ CHANGELOG.md
@@ -9,10 +9,12 @@
 ### New Features
 - Support setting PyTorch's device with thye `--device` flag (@gregwar)
 - Added `vec_env_wrapper` support in the config (works the same as `env_wrapper`)
+- Added Huggingface hub integration
 
 ### Bug fixes
 - Fix `Reacher-v3` name in PPO hyperparameter file
 - Pinned ale-py==0.7.4 until new SB3 version is released
+- Fix enjoy / record videos with LSTM policy
 
 ### Documentation