+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### New Features
 - Support setting PyTorch's device with thye `--device` flag (@gregwar)
-- Add `--tital-n-trials` parameter to help with distributed optimization. (@ernestum)
+- Add `--total-n-trials` parameter to help with distributed optimization. (@ernestum)
 - Added `vec_env_wrapper` support in the config (works the same as `env_wrapper`)
 - Added Huggingface hub integration